+++ celldetective/gui/classifier_widget.py
@@ -13,8 +13,8 @@ from PyQt5.QtWidgets import (
 )
 from PyQt5.QtCore import Qt, QSize
 from PyQt5.QtGui import QCloseEvent
-from superqt import QLabeledSlider, QSearchableComboBox
-from celldetective.gui.base.sliders import QLabeledDoubleSlider
+from superqt import QSearchableComboBox
+from celldetective.gui.base.sliders import QLabeledSlider, QLabeledDoubleSlider
 from superqt.fonticon import icon
 from fonticon_mdi6 import MDI6