+++ celldetective/gui/generic_signal_plot.py
@@ -21,8 +21,7 @@ from celldetective.gui.tableUI import TableUI
 from celldetective.utils.experiment import collect_experiment_metadata
 
 from superqt.fonticon import icon
-from superqt import QLabeledSlider
-from celldetective.gui.base.sliders import QLabeledDoubleSlider
+from celldetective.gui.base.sliders import QLabeledSlider, QLabeledDoubleSlider
 from fonticon_mdi6 import MDI6
 import numpy as np
 import json