+++ celldetective/gui/settings/_settings_segmentation_model_training.py
@@ -22,8 +22,7 @@ from celldetective.gui.base.channel_norm_generator import ChannelNormGenerator
 import multiprocessing
 from celldetective.gui.workers import Runner
 from celldetective.gui.dynamic_progress import DynamicProgressDialog
-from superqt import QLabeledSlider
-from celldetective.gui.base.sliders import QLabeledDoubleSlider
+from celldetective.gui.base.sliders import QLabeledSlider, QLabeledDoubleSlider
 from superqt.fonticon import icon
 from fonticon_mdi6 import MDI6
 from celldetective.utils.model_getters import (