+++ celldetective/gui/settings/_settings_tracking.py
@@ -29,8 +29,7 @@ from celldetective.gui.base.list_widget import ListWidget
 from celldetective.gui.base.feature_choice import FeatureChoice
 from celldetective.gui.base.components import QHSeperationLine
 from celldetective.gui.base.collapsible import CollapsibleFrame
-from superqt import QLabeledSlider
-from celldetective.gui.base.sliders import QLabeledDoubleSlider
+from celldetective.gui.base.sliders import QLabeledSlider, QLabeledDoubleSlider
 from superqt.fonticon import icon
 from fonticon_mdi6 import MDI6
 from celldetective import get_software_location